added NotificationSettings to WelcomeScreen.js and now asking for permissions if necessary
